Initial commit of pubmed functions
Search function currently accepts either a search query (str), a single id (int), or a list of ids.  For search queries, it uses the pubmed request functionality to retrieve some number of IDs related to the search.  For now, it just passes the list of ids back out.  Next step is to use the efetch or esummary tool to retreive desired information.